backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.104
botocore==1.29.104
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
decorator==4.4.2
deepgram-sdk==2.3.0
Flask==2.2.3
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
premailer==3.10.0
proglog==0.1.10
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
SpeechRecognition==3.10.0
tqdm==4.65.0
urllib3==1.26.15
websockets==10.4
Werkzeug==2.2.3
yagmail==0.15.293
yarl==1.8.2
zipp==3.15.0
